Add support for multiple keyspaces
Making AstyanaxClient lowercase the keyspace.  It seems to have trouble with an uppercase keysapce.  Might be a Cassandra 1.2 thing.

Making only the lookup lowercase and not the actual keyspace.  This is to support Cassandra connections before 1.2.

Updating examples to handle more than one keyspace.

Updating examples to handle more than one keyspace. (Missed a Topology config)